requirements.txt
autogen-agentchat==0.4.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.111.0
uvicorn==0.30.0
colorlog==6.8.2
streamlit==1.35.0
plotly==5.22.0
pydantic==2.7.0
python-multipart==0.0.9